Fixed #22341 -- Turned django.db.models.fields.related into a package.
