+++ django/db/models/fields/related/__init__.py
@@ -1,0 +1,31 @@
+from django.db.models.fields.related.base import (  # NOQA
+    RECURSIVE_RELATIONSHIP_CONSTANT,
+    add_lazy_relation,
+    do_pending_lookups,
+    RelatedField,
+    ReverseSingleRelatedObjectDescriptor,
+    create_foreign_related_manager,
+    ForeignRelatedObjectsDescriptor,
+    ForeignObjectRel,
+    ForeignObject,
+)
+
+from django.db.models.fields.related.many_to_many import (  # NOQA
+    create_many_to_many_intermediary_model,
+    create_many_related_manager,
+    ManyRelatedObjectsDescriptor,
+    ReverseManyRelatedObjectsDescriptor,
+    ManyToManyRel,
+    ManyToManyField,
+)
+
+from django.db.models.fields.related.many_to_one import (  # NOQA
+    ManyToOneRel,
+    ForeignKey,
+)
+
+from django.db.models.fields.related.one_to_one import (  # NOQA
+    SingleRelatedObjectDescriptor,
+    OneToOneRel,
+    OneToOneField,
+)
